requirements.txt
docling==2.9.0
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.19
python-decouple==3.8


